refactor: address review
