+++ python/beeai_framework/agents/types.py
@@ -14,8 +14,8 @@
 
 from pydantic import BaseModel, InstanceOf
 
-from beeai_framework.cancellation import AbortSignal
 from beeai_framework.tools.tool import AnyTool
+from beeai_framework.utils import AbortSignal
 
 
 class BaseAgentRunOptions(BaseModel):
