+++ python/beeai_framework/errors.py
@@ -158,30 +158,4 @@ class AbortError(FrameworkError, CancelledError):
         super().__init__(message, is_fatal=True, is_retryable=False, cause=cause, context=context)
 
 
-class LoggerError(FrameworkError):
-    """Raised for errors caused by logging."""
-
-    def __init__(
-        self,
-        message: str = "Logger error",
-        *,
-        cause: Exception | None = None,
-        context: dict[str, Any] | None = None,
-    ) -> None:
-        super().__init__(message, is_fatal=True, is_retryable=False, cause=cause, context=context)
-
-
-class PromptTemplateError(FrameworkError):
-    """Raised for errors caused by PromptTemplate."""
-
-    def __init__(
-        self,
-        message: str = "PromptTemplate error",
-        *,
-        cause: Exception | None = None,
-        context: dict[str, Any] | None = None,
-    ) -> None:
-        super().__init__(message, is_fatal=True, is_retryable=False, cause=cause, context=context)
-
-
-__all__ = ["AbortError", "FrameworkError", "LoggerError", "PromptTemplateError"]
+__all__ = ["AbortError", "FrameworkError"]
