+++ python/beeai_framework/tools/code/sandbox.py
@@ -18,10 +18,10 @@ from typing import Any, Self
 from pydantic import BaseModel
 
 from beeai_framework.context import RunContext
-from beeai_framework.emitter.emitter import Emitter
+from beeai_framework.emitter import Emitter
 from beeai_framework.errors import FrameworkError
+from beeai_framework.tools import Tool
 from beeai_framework.tools.code.python import PythonTool
-from beeai_framework.tools.tool import Tool
 from beeai_framework.tools.types import StringToolOutput, ToolRunOptions
 from beeai_framework.utils.models import JSONSchemaModel
 
@@ -46,7 +46,7 @@ class SandboxToolOptions(ToolRunOptions):
 class SandboxTool(Tool[BaseModel, SandboxToolOptions, StringToolOutput]):
     def __init__(self, options: SandboxToolOptions) -> None:
         super().__init__()
-        self.__options = options
+        self._tool_options = options
 
     def _create_emitter(self) -> Emitter:
         return Emitter.root().child(
@@ -56,28 +56,28 @@ class SandboxTool(Tool[BaseModel, SandboxToolOptions, StringToolOutput]):
 
     @property
     def name(self) -> str:
-        return self.__options.name
+        return self._tool_options.name
 
     @property
     def description(self) -> str:
-        return self.__options.description
+        return self._tool_options.description
 
     @property
     def input_schema(self) -> type[BaseModel]:
-        return JSONSchemaModel.create(self.name, self.__options.input_schema)
+        return JSONSchemaModel.create(self.name, self._tool_options.input_schema)
 
     async def _run(
         self, tool_input: BaseModel | dict[str, Any], options: SandboxToolOptions | None, context: RunContext
     ) -> StringToolOutput:
         try:
             result = await PythonTool.call_code_interpreter(
-                f"{self.__options.code_interpreter_url}/v1/execute-custom-tool",
+                f"{self._tool_options.code_interpreter_url}/v1/execute-custom-tool",
                 {
-                    "tool_source_code": self.__options.source_code,
+                    "tool_source_code": self._tool_options.source_code,
                     "tool_input_json": tool_input.model_dump_json()
                     if isinstance(tool_input, BaseModel)
                     else json.dumps(tool_input),
-                    "env": {**self.__options.env, **(options.env if options else {})},
+                    "env": {**self._tool_options.env, **(options.env if options else {})},
                 },
             )
