+++ python/beeai_framework/tools/search/duckduckgo.py
@@ -19,7 +19,7 @@ from pydantic import BaseModel, Field
 from beeai_framework.context import RunContext
 from beeai_framework.emitter.emitter import Emitter
 from beeai_framework.logger import Logger
-from beeai_framework.tools.errors import ToolError
+from beeai_framework.tools import ToolError
 from beeai_framework.tools.search import SearchToolOutput, SearchToolResult
 from beeai_framework.tools.tool import Tool
 from beeai_framework.tools.types import ToolRunOptions
