+++ python/beeai_framework/tools/types.py
@@ -17,7 +17,7 @@ from typing import Any
 
 from pydantic import BaseModel
 
-from beeai_framework.cancellation import AbortSignal
+from beeai_framework.utils import AbortSignal
 from beeai_framework.utils.strings import to_json
