+++ python/beeai_framework/utils/__init__.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
+from beeai_framework.utils.cancellation import AbortController, AbortSignal
 from beeai_framework.utils.config import CONFIG
 from beeai_framework.utils.models import JSONSchemaModel, ModelLike
 from beeai_framework.utils.types import MaybeAsync
 
-__all__ = ["CONFIG", "JSONSchemaModel", "MaybeAsync", "ModelLike"]
+__all__ = ["CONFIG", "AbortController", "AbortSignal", "JSONSchemaModel", "MaybeAsync", "ModelLike"]
