+++ python/beeai_framework/utils/cancellation.py
@@ -110,6 +110,3 @@ async def abort_signal_handler(
     finally:
         if signal:
             signal.remove_event_listener(abort_handler)
-
-
-__all__ = ["AbortSignal"]
