+++ python/examples/agents/react_advanced.py
@@ -9,7 +9,6 @@ from pydantic import BaseModel
 from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.agents import AgentExecutionConfig
 from beeai_framework.agents.react import ReActAgent
-from beeai_framework.cancellation import AbortSignal
 from beeai_framework.emitter import Emitter, EmitterOptions, EventMeta
 from beeai_framework.errors import FrameworkError
 from beeai_framework.memory import UnconstrainedMemory
@@ -17,6 +16,7 @@ from beeai_framework.template import PromptTemplateInput
 from beeai_framework.tools import AnyTool
 from beeai_framework.tools.search import DuckDuckGoSearchTool
 from beeai_framework.tools.weather import OpenMeteoTool
+from beeai_framework.utils import AbortSignal
 from examples.helpers.io import ConsoleReader
 
 # Load environment variables
