+++ python/examples/backend/providers/anthropic.py
@@ -5,11 +5,11 @@ from pydantic import BaseModel, Field
 
 from beeai_framework.adapters.anthropic import AnthropicChatModel
 from beeai_framework.backend import ChatModel, ChatModelNewTokenEvent, UserMessage
-from beeai_framework.cancellation import AbortSignal
 from beeai_framework.emitter import EventMeta
 from beeai_framework.errors import AbortError
 from beeai_framework.parsers.field import ParserField
 from beeai_framework.parsers.line_prefix import LinePrefixParser, LinePrefixParserNode
+from beeai_framework.utils import AbortSignal
 
 MODEL_NAME: Final[str] = "claude-3-haiku-20240307"
