+++ python/tests/templates/test_templates.py
@@ -20,9 +20,9 @@ from zoneinfo import ZoneInfo
 import pytest
 from pydantic import BaseModel, ValidationError
 
-from beeai_framework.errors import PromptTemplateError
 from beeai_framework.template import (
     PromptTemplate,
+    PromptTemplateError,
     PromptTemplateInput,
 )
